% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_check_def_playing_styles.R
\name{do_check_def_playing_styles}
\alias{do_check_def_playing_styles}
\title{Check the definition of PES Master playing styles}
\usage{
do_check_def_playing_styles(pesmaster_playing_styles, pesmaster_data, 
                            pesmaster_algorithm, type_player, rank_value = 50, 
                            skip_grams_window_value = 5, x_max_value = 10, 
                            alpha = 0.7, top_simil = 3, max_scaling = FALSE,
                            verbose)
}
\arguments{
\item{pesmaster_playing_styles}{PES Master playing styles.}

\item{pesmaster_data}{PES Master data.}

\item{pesmaster_algorithm}{Average values for all the variables. This is computed
with \code{do_prepare_profiles}.}

\item{type_player}{String with the position. Options are Goalkeeper, Defender, Midfielder and Forward.}

\item{rank_value}{Number to indicate the desired dimension for the latent vectors. Default 50.}

\item{skip_grams_window_value}{Number to indicate the window for term-co-occurrence 
matrix construction. Default 5.}

\item{x_max_value}{Number to indicate the maximum number of co-occurrences to use in the 
weighting function. Default 10.}

\item{alpha}{Number between 0 and 1 to combine text-based cosine similarity 
and numeric similarity into a final similarity score. The closer to 0, the more important 
are the numeric variables. Default 0.7.}

\item{top_simil}{Number of potential candidates to return.}

\item{max_scaling}{Logical to indicate if the term-co-occurrence matrix must be scaled. 
This is suggested when there are words that are repeated in most documents. Default FALSE.}

\item{verbose}{Should R report information on progress? TRUE or FALSE.}
}
\value{
A plot with a kind of confusion matrix.
}
\description{
This is an experiment to check how the PES Master playing styles are defined.
The experiment consists of sending a query with the playing style text
description and the corresponding average numeric variables to the player
recruiting algorithm and check how many times the target playing style is
returned.
}
\examples{
\dontrun{
# Prepare the data with the abilities related to each position:
pesmaster_pos_abi <- data.frame(type = rep("Goalkeeper", 6),
                                ability = c("goalkeeping", "gk_awareness", "gk_catching", 
                                            "gk_parrying", "gk_reflexes", "gk_reach" ))
                                            
pesmaster_algorithm <- do_prepare_profiles(pesmaster_sample, 
                                           pesmaster_playing_styles, 
                                           pesmaster_pos_abi)

do_check_def_playing_styles(pesmaster_playing_styles, pesmaster_sample, pesmaster_algorithm, 
                            c("Defensive Goalkeeper", "Offensive Goalkeeper"), 50, 5, 10, 
                            0.5, 10, TRUE, TRUE)
}                           

}
\author{
Guillermo Vinue
}
