% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_explore_grid_param.R
\name{do_explore_grid_param}
\alias{do_explore_grid_param}
\title{Explore Hyperparameter Grid Performance}
\usage{
do_explore_grid_param(grid_param)
}
\arguments{
\item{grid_param}{Data set that includes different combinations of 
hyperparameters and the corresponding sensitivity values.}
}
\value{
A plot.
}
\description{
Visualize the sensitivity values for the performance of the player 
recruitment algorithm using a grid of different combinations of
hyperparameters.
}
\examples{
\dontrun{
do_explore_grid_param(grid_param_fawir)
}                           

}
\author{
Guillermo Vinue
}
