% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_explore_leagues.R
\name{do_explore_leagues}
\alias{do_explore_leagues}
\title{Explore PES Master leagues}
\usage{
do_explore_leagues(pesmaster_data)
}
\arguments{
\item{pesmaster_data}{PES Master data.}
}
\value{
A plot.
}
\description{
Visualize the number of teams and players per league.
}
\examples{
\dontrun{
do_explore_leagues(pesmaster_sample)
}                           

}
\author{
Guillermo Vinue
}
