% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_explore_playing_styles.R
\name{do_explore_playing_styles}
\alias{do_explore_playing_styles}
\title{Explore PES Master playing styles}
\usage{
do_explore_playing_styles(pesmaster_data)
}
\arguments{
\item{pesmaster_data}{PES Master data.}
}
\value{
A plot.
}
\description{
Visualize the number of playing styles.
}
\examples{
\dontrun{
do_explore_playing_styles(pesmaster_sample)
}                           

}
\author{
Guillermo Vinue
}
