% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_explore_profiles.R
\name{do_explore_profiles}
\alias{do_explore_profiles}
\title{Explore PES Master player profiles}
\usage{
do_explore_profiles(pes_algorithm, type_player)
}
\arguments{
\item{pes_algorithm}{Average values for all the variables. This is computed
with \code{do_prepare_profiles}.}

\item{type_player}{String with the position. Options are Goalkeeper,
Defender, Midfielder and Forward.}
}
\value{
A plot.
}
\description{
Visualize the average values of each variable for a certain type of player.
This is an interesting way to establish player profiles.
}
\examples{
\dontrun{
# Prepare the data with the abilities related to each position:
pesmaster_pos_abi <- data.frame(type = rep("Goalkeeper", 6),
                                ability = c("goalkeeping", "gk_awareness", "gk_catching", 
                                            "gk_parrying", "gk_reflexes", "gk_reach" ))
                                            
pesmaster_algorithm <- do_prepare_profiles(pesmaster_sample, 
                                           pesmaster_playing_styles, 
                                           pesmaster_pos_abi)                                             
                                            
do_explore_profiles(pesmaster_algorithm, "Goalkeeper")
}                           

}
\author{
Guillermo Vinue
}
