% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_explore_texts.R
\name{do_explore_texts}
\alias{do_explore_texts}
\title{Explore PES Master texts}
\usage{
do_explore_texts(pesmaster_data, pesmaster_positions, custom_stopwords)
}
\arguments{
\item{pesmaster_data}{PES Master data.}

\item{pesmaster_positions}{PES Master playing positions.}

\item{custom_stopwords}{Custom stop words to add to the default lists.}
}
\value{
A plot.
}
\description{
Visualize the words that best describe each playing position.
}
\examples{
\dontrun{
library(tibble)
mystopwords <- tibble(word = c("goalkeeper", "goalkeepers", "player", "ball", "frequently"))
do_explore_texts(pesmaster_sample, pesmaster_positions, mystopwords)
}                           

}
\author{
Guillermo Vinue
}
