% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_prepare_profiles.R
\name{do_prepare_profiles}
\alias{do_prepare_profiles}
\title{Prepare PES Master player profiles}
\usage{
do_prepare_profiles(pesmaster_data, pesmaster_playing_styles, pesmaster_pos_abi)
}
\arguments{
\item{pesmaster_data}{PES Master data.}

\item{pesmaster_playing_styles}{PES Master playing styles.}

\item{pesmaster_pos_abi}{PES Master abilities related to each position.}
}
\value{
A data frame.
}
\description{
Prepare the player profiles by computing the mean values of all 
the variables related to the same type of player.
}
\examples{
\dontrun{
# Prepare the data with the abilities related to each position:
pesmaster_pos_abi <- data.frame(type = rep("Goalkeeper", 6),
                                ability = c("goalkeeping", "gk_awareness", "gk_catching", 
                                            "gk_parrying", "gk_reflexes", "gk_reach" ))
                                            
pesmaster_algorithm <- do_prepare_profiles(pesmaster_sample, 
                                           pesmaster_playing_styles, 
                                           pesmaster_pos_abi)
}                           

}
\author{
Guillermo Vinue
}
