\name{grid_param_fawir}
\alias{grid_param_fawir}
\docType{data}
\title{
Example grid of hyperparameters
}
\description{
Sample of the data set that includes different combinations of 
hyperparameters and the corresponding sensitivity values.
}
\usage{grid_param_fawir}
\format{Data frame with 90 rows and 5 columns.}
\keyword{datasets}
