\name{pesmaster_playing_styles}
\alias{pesmaster_playing_styles}
\docType{data}
\title{
PES Master playing styles
}
\description{
A data frame with three columns: the general category, the related playing style
and the text description.
}
\usage{pesmaster_playing_styles}
\format{Data frame with 22 rows and 3 columns.}
\source{\url{https://www.pesmaster.com/efootball-2022/}}
\keyword{datasets}
