\name{pesmaster_positions}
\alias{pesmaster_positions}
\docType{data}
\title{
PES Master players' positions
}
\description{
A data frame with three columns: the acronym for each position, the corresponding
explanation and the general category.
}
\usage{pesmaster_positions}
\format{Data frame with 13 rows and 3 columns.}
\source{\url{https://www.pesmaster.com/efootball-2022/}}
\keyword{datasets}
