\name{pesmaster_sample}
\alias{pesmaster_sample}
\docType{data}
\title{
PES Master data sample
}
\description{
Sample of data obtained from the PES Master database for illustration purposes.
There are both numeric and text variables.
}
\usage{pesmaster_sample}
\format{Data frame with 63 rows and 41 columns.}
\source{\url{https://www.pesmaster.com/efootball-2022/}}
\keyword{datasets}
